colab-requirements.txt
absl-py==2.3.1
astunparse==1.6.3
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.4.0
google-auth==2.41.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
joblib==1.5.2
keras-preprocessing==1.1.2
kiwisolver==1.4.9
libclang==18.1.1
lz4==4.4.4
markdown==3.9
markupsafe==3.0.3
matplotlib==3.10.7
mtcnn==1.0.0
numpy==1.26.4
oauthlib==3.3.1
opencv-contrib-python==4.12.0.88
opt-einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
werkzeug==3.1.3
wrapt==2.0.0

# --- API (Backend) ---
fastapi
uvicorn[standard]

# --- Frontend ---
streamlit
requests